fix: coerce empty donated-item price fields to 0.0 to avoid 500 on save
Blank SellPrice/EstPrice/MaterialValue/MinimumPrice inputs arrived as ''
and were passed straight to the Propel DECIMAL setters, which MySQL
rejects (Incorrect decimal value: '' for column di_sellprice) -> 500.
Cast the four money fields to (float) so blanks become 0.0. Covers both
the create and update branches, which share these variables.

Added a Cypress regression asserting a donated-item save with empty
price fields returns 302 (not 500).

+++ src/fundraiser/routes/donated-item.php
@@ -158,10 +158,11 @@ $app->post('/{fundraiserId}/donated-items/editor[/{itemId}]', function (Request 
     $iBuyer        = (int) InputUtils::legacyFilterInput($body['Buyer'] ?? '0', 'int');
     $sTitle        = InputUtils::legacyFilterInput($body['Title'] ?? '');
     $sDescription  = InputUtils::legacyFilterInput($body['Description'] ?? '');
-    $nSellPrice    = InputUtils::legacyFilterInput($body['SellPrice'] ?? '');
-    $nEstPrice     = InputUtils::legacyFilterInput($body['EstPrice'] ?? '');
-    $nMaterialValue = InputUtils::legacyFilterInput($body['MaterialValue'] ?? '');
-    $nMinimumPrice = InputUtils::legacyFilterInput($body['MinimumPrice'] ?? '');
+    // DECIMAL columns reject empty/non-numeric strings, so coerce blank price fields to 0.0.
+    $nSellPrice     = (float) InputUtils::legacyFilterInput($body['SellPrice'] ?? '');
+    $nEstPrice      = (float) InputUtils::legacyFilterInput($body['EstPrice'] ?? '');
+    $nMaterialValue = (float) InputUtils::legacyFilterInput($body['MaterialValue'] ?? '');
+    $nMinimumPrice  = (float) InputUtils::legacyFilterInput($body['MinimumPrice'] ?? '');
     $sPictureURL   = InputUtils::legacyFilterInput($body['PictureURL'] ?? '');
 
     if (!$bMultibuy) { $bMultibuy = 0; }
